python-otp-gen/requirements.txt
pyotp>=2.9.0
qrcode>=7.4.2
Pillow>=10.0.0

